Refactored to resolve a circular dependency
